add upgrade recipe
Even though the newest (as of right now) release is set to install by
default, the attributes are set to 'upgrade' for splunk 4.3.7. This is
because internally, we use 4.2.1 and need to upgrade to 4.3.x before
we can upgrade to 6.0.1. Our internal splunk cookbook for 4.2.1 will
be used alongside this one until the upgrade is performed. This is
probably not a common use case, but the pattern is available for those
who want to reuse it. Simply set the appropriate attributes (see
`attributes/upgrade.rb`)