Refactor the setup_shclustering recipe
* remove resource for executing the `init-shcluster-config` command to
set our shcluster configs
* remove the sections for building up the command params that were used
in that command since they are no longer needed
* change the setting of `shcluster_secret` so that it can either be
passed in via chef-vault or assigned in clear text via
`node['splunk']['shclustering']['shcluster_secret’]`. This secret gets
written to the server.conf in clear text, but is then encrypted when
Splunk restarts.
* change the setting of `shcluster_servers_list` so that it can either
be statically assigned via the
`node['splunk']['shclustering']['shcluster_members’]` array or
dynamically populated by a Chef search for other nodes that are:
** in the same environment
** have search head clustering enabled
** and have the same shcluster label set
* implement the setting shcluster configuration by creating a custom
splunk app with a server.conf that is templated out using the shcluster
config attributes we pass in.